requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
Werkzeug
PyMySQL
